Fjern original brukersystem til fordel for feide. Fjern mulighet for redigering av navn/epost. Misc rydding.
