--- authentication/templates/authentication/forgot_password.html
@@ -1,30 +1,0 @@
-{% extends 'website/base.html' %}
-{% load staticfiles %}
-{% block content %}
-<div class="section block hs-green white-text">
-	<div class="container">
-		<div class="row">
-			<div class="col s12">
-				<h3>Glemt passord</h3>
-			</div>
-		</div>
-	</div>
-</div>
-<div class="section block">
-	<div class="container">
-		<form action="/authentication/forgot-password/" method="post">
-			{% csrf_token %}
-			<div class="row">
-				<div class="input-field col push-m2 pull-m2 s12 m8">
-					<p>{{ form.email.errors }}</p>
-					{{ form.email }}
-					<label for="{{ form.email.id_for_label }}">Email</label>
-				</div>
-				<div class="col push-m2 pull-m2 s12 m8">
-					<button id="sendNewPasswordButton" class="btn waves-effect waves-light" type="submit" name="action">Send</button>
-				</div>
-			</div>
-		</form>
-	</div>
-</div>
-{% endblock content %}
