--- authentication/templates/authentication/login_modal.html
@@ -1,29 +1,0 @@
-<form action="/authentication/login/" method="post">
-	{% csrf_token %}
-			<div class="row center">
-				<i class="material-icons large hs-gray-text">account_circle</i>
-			</div>
-			<form action="/authentication/login/" method="post">
-				<div class="row">
-					<div class="input-field col s12">
-						{{ login_form.username }}
-						<label for="{{ login_form.username.id_for_label }}">{{ login_form.username.label }}</label>
-					</div>
-					<p>{{ login_form.username.errors }}</p>
-					<div class="input-field col s12">
-						{{ login_form.password }}
-						<label for="{{ login_form.password.id_for_label }}">{{ login_form.password.label }}</label>
-					</div>
-					<p>{{ login_form.password.errors }}</p>
-				</div>
-				<div class="row">
-					<div class="input-field col s12">
-						<button class="btn btn-large waves-effect waves-light hs-green" type="submit" name="action">Logg inn</button>
-						<a class="right" href="/authentication/forgot-password">Glemt passord?</a>
-					</div>
-					<div class="col s12"> 
-						<a class="btn btn-large waves-effect waves-light white-text hs-green" href="{% url 'authentication_feide:login' %}">Logg inn med FEIDE</a>
-					</div>
-				</div>
-			</form>
-</form>
