--- authentication/templates/authentication/password_reset_email.html
@@ -1,14 +1,0 @@
-{% load i18n %}{% autoescape off %}
-{% blocktrans %}You're receiving this email because you requested a password reset for your user account at Hackerspace.{% endblocktrans %}
-
-{% trans "Please go to the following page and choose a new password:" %}
-{% block reset_link %}
-{{ protocol }}://hackerspace-ntnu.no/authentication/reset/{{ uid }}/{{ token }}
-{% endblock %}
-{% trans "Your login email, in case you've forgotten, is same this email address:" %} {{ user.get_username }}
-
-{% trans "Thanks for using our site!" %}
-
-{% blocktrans %}The DevOps team{% endblocktrans %}
-
-{% endautoescape %}
