--- authentication/templates/authentication/redirection_page.html
@@ -1,20 +1,0 @@
-{% extends 'website/base.html' %}
-{% load staticfiles %}
-
-{% block header %}
-  <meta http-equiv="refresh" content="3;url=/">
-{% endblock header %}
-
-{% block content %}
-    <main>
-    <div class="section no-pad-bot" id="index-banner">
-        <div class="container">
-            <div class="row">
-                <div>
-                    <h3 class="center-align" style="font-weight:300;">{{ title }}</h3>
-                </div>
-            </div>
-        </div>
-    </div>
-    </main>
-{% endblock %}
