--- authentication/templates/authentication/set_password.html
@@ -1,39 +1,0 @@
-{% extends 'website/base.html' %}
-{% load staticfiles %}
-
-{% block header %}
-{% endblock header %}
-
-{% block content %}
-<div class="section block hs-green white-text">
-	<div class="container">
-		<div class="row">
-			<div class="col s12">
-				<h3>Nytt passord</h3>
-			</div>
-		</div>
-	</div>
-</div>
-<div class="section block">
-	<div class="container">
-		<form action="" method="post">
-			{% csrf_token %}
-			<div class="row">
-				<div class="input-field col s12">
-					<p>{{ form.new_password1.errors }}</p>
-					{{ form.new_password1 }}
-					<label for="{{ form.new_password1.id_for_label }}">Passord</label>
-				</div>
-				<div class="input-field col s12">
-					<p>{{ form.new_password2.errors }}</p>
-					{{ form.new_password2 }}
-					<label for="{{ form.new_password2.id_for_label }}">Bekreft passord</label>
-				</div>
-				<div class="col s12">
-					<button class="btn waves-effect waves-light" type="submit" name="action">Bytt passord</button>
-				</div>
-			</div>
-		</form>
-	</div>
-</div>
-{% endblock content %}
