--- authentication/templates/authentication/signup.html
@@ -1,74 +1,0 @@
-{% extends 'website/base.html' %}
-{% load staticfiles %}
-{% block content %}
-<div class="section block hs-green white-text">
-	<div class="container">
-		<div class="row">
-			<div class="col s12">
-				<h3>Registering av brukerkonto</h3>
-				<div><a href="{% url 'authentication_feide:login' %}"><button  class="btn waves-effect waves-light white grey-text">Logg inn med FEIDE</button></a></div>
-			</div>
-		</div>
-	</div>
-</div>
-<div class="section block">
-	<div class="container">
-		<div class="card-panel">
-		<form action="/authentication/signup/" method="post">
-			{% csrf_token %}
-			<div class="row">
-				<div class="input-field col s12">
-					<i class="material-icons prefix">account_circle</i>
-					{{ form.username }}
-					<label for="{{ form.username.id_for_label }}">{{ form.username.label }}</label>
-					{% for error in form.username.errors %}
-					<span class="helper-text" data-error="{{ error | escape }}"></span>
-					{% endfor %}
-				</div>
-				<div class="input-field col s6">
-					<i class="material-icons prefix">perm_identity</i>
-					{{ form.first_name }}
-					<label for="{{ form.first_name.id_for_label }}">{{ form.first_name.label }}</label>
-					<span class="helper-text" data-error="{{ form.first_name.errors | striptags }}"></span>
-				</div>
-				<div class="input-field col s6">
-					{{ form.last_name }}
-					<label for="{{ form.last_name.id_for_label }}" data-error="{{ form.last_name.errors | striptags }}"> {{ form.last_name.label }} </label>
-				</div>
-				<div class="input-field col s12">
-					<i class="material-icons prefix">email</i>
-					{{ form.email }}
-					<label for="{{ form.email.id_for_label }}"> {{ form.email.label }} </label>
-					<span class="helper-text" data-error="{{ form.email.errors | striptags }}"></span>
-				</div>
-				<div class="input-field col s6">
-					<i class="material-icons prefix">vpn_key</i>
-					{{ form.password1 }}
-					<label for="{{ form.password1.id_for_label }}">{{ form.password1.label }}</label> 
-					<span class="helper-text" data-error="{{ form.password1.errors | striptags }}"></span>
-				</div>
-				<div class="input-field col s6">
-					{{ form.password2 }}
-					<label for="{{ form.password2.id_for_label }}"> {{ form.password2.label }} </label>
-					<span class="helper-text active" data-error="{{ form.password2.errors | striptags }}"></span>
-				</div>
-				<div class="col s12">
-					<p>
-						<label>
-							{{ form.tos_accept }}
-							<span>{{ form.tos_accept.label }}</span>
-						</label>
-					<a target="_blank" href="/tos">her</a>
-					</p>
-				</div>
-				<br>
-				<br>
-				<div class="col s12">
-					<button id="signUpButton" class="btn btn-large waves-effect waves-light hs-green" type="submit" name="action">Registrer</button>
-				</div>
-			</div>
-		</form>
-		</div>
-	</div>
-</div>
-{% endblock content %}
