--- authentication/templates/authentication/signup_mail.html
@@ -1,8 +1,0 @@
-<p>Velkommen som bruker på Hackerspace!</p>
-
-<a href="http://hackerspace-ntnu.no/authentication/signup-confirm/{{ uid }}/{{ token }}/">Trykk for å aktivere konto</a>
-</p>
-
-
-<p>Mvh, DevOps-gruppen på Hackerspace</p>
-
