+++ userprofile/admin.py
@@ -1,5 +1,5 @@
 from django.contrib import admin
-from .models import Profile, Skill, DutyTime, Group
+from .models import Profile, Skill, Group
 
 
 class ProfileAdmin(admin.ModelAdmin):
@@ -10,5 +10,4 @@ class ProfileAdmin(admin.ModelAdmin):
 
 admin.site.register(Profile, ProfileAdmin)
 admin.site.register(Skill)
-admin.site.register(DutyTime)
 admin.site.register(Group)
