+++ userprofile/models.py
@@ -4,7 +4,6 @@ from django.core.files.base import ContentFile
 from django.dispatch import receiver
 from django.db import models
 from django.shortcuts import reverse
-
 from datetime import datetime
 from sorl.thumbnail import get_thumbnail
 
@@ -33,38 +32,6 @@ class Group(models.Model):
         return self.title
 
 
-class DutyTime(models.Model):
-    MONDAY = 'Mandag'
-    TUESDAY = 'Tirsdag'
-    WEDNESDAY = 'Onsdag'
-    THURSDAY = 'Torsdag'
-    FRIDAY = 'Fredag'
-
-    DUTYDAYS_CHOICES = (
-        (MONDAY, 'Mandag'),
-        (TUESDAY, 'Tirsdag'),
-        (WEDNESDAY, 'Onsdag'),
-        (THURSDAY, 'Torsdag'),
-        (FRIDAY, 'Fredag')
-    )
-
-    day = models.CharField(max_length=7, choices=DUTYDAYS_CHOICES, default=MONDAY)
-    start_time = models.TimeField()
-    end_time = models.TimeField()
-
-    def shorten_time(self, time):
-        return ":".join(str(time).split(":")[:2])
-
-    def dutyday(self):
-        return [dutyday[1] for dutyday in self.DUTYDAYS_CHOICES if dutyday[0] == self.day][0]
-
-    def dutytime(self):
-        return self.shorten_time(self.start_time) + " - " + self.shorten_time(self.end_time)
-
-    def __str__(self):
-        return self.day + " " + str(self.start_time) + "-" + str(self.end_time)
-
-
 class Profile(models.Model):
     user = models.OneToOneField(User, related_name='profile', on_delete=models.CASCADE)
     group = models.ManyToManyField(Group, related_name='profile', verbose_name="Gruppe", blank=True)
@@ -73,9 +40,6 @@ class Profile(models.Model):
     access_card = models.CharField(max_length=20, null=True, blank=True)
     study = models.CharField(max_length=50, null=True, blank=True)
     skills = models.ManyToManyField(Skill, related_name="skills", blank=True)
-    duty = models.ManyToManyField(DutyTime, related_name="duty", blank=True)
-
-    auto_duty = models.BooleanField(default=True)
     tos_accepted = models.BooleanField(default=False)
 
     def save(self, *args, **kwargs):
@@ -91,10 +55,3 @@ class Profile(models.Model):
 
     def get_absolute_url(self):
         return reverse('userprofile:profile', args=(self.pk,))
-
-
-# Save a user profile whenever we create a user
-@receiver(post_save, sender=User, dispatch_uid="create_profile_on_user_create")
-def create_profile(sender, instance, created, **kwargs):
-    if created:
-        Profile.objects.create(user=instance, auto_duty=False, tos_accepted=True)
