+++ userprofile/templates/userprofile/edit_profile.html
@@ -6,7 +6,7 @@
 	<div class="container">
 		<div class="row">
 			<div class="col s12 white-text">
-				<h4>Innsillinger for {{ object.get_full_name }}</h4>
+				<h4>Innsillinger for {{ profile.user.get_full_name }}</h4>
 			</div>
 		</div>
 	</div>
@@ -23,46 +23,20 @@
 			</div>
 		</div>
 		<form enctype="multipart/form-data" method="post" action="">
-			{{ formset.management_form }}
 			{% csrf_token %}
-			<div class="row">
-				<div class="col s12">
-					<h5>Generell kontoinformasjon</h5>
-					<div class="divider"></div>
-					<br>
-				</div>
-				<div class="col s6 input-field">
-					{{ form.first_name }}
-					<label for="{{form.first_name.id_for_label }}">{{ form.first_name.label }}</input>
-				</div>
-				<div class="col s6 input-field">
-					{{ form.last_name }}
-					<label for="{{form.last_name.id_for_label }}">{{ form.last_name.label }}</input>
-				</div>
-				<div class="col s12 input-field">
-					{{ form.email }}
-					<label for="{{form.email.id_for_label }}">{{ form.email.label }}</input>
-				</div>
-				<div class="col s12">
-					<a target="_blank" class="btn btn-flat hs-green white-text" href="{% url 'auth:password_change' %}">Endre passord</a>
-				</div>
-			</div>
-
 			<div class="row">
 				<div class="col s12">
 					<h5>Profil</h5>
 					<div class="divider"></div>
 					<br>
 				</div>
-				{% for form in formset %}
-					{{ form.errors }}
 					{% for field in form.visible_fields %}
+						{{ field.errors }}
 						{% if field.label == "Profilbilde" %}
 						<div class="col s12 file-field input-field">
 							<div class="btn hs-green white-text">
-								<span>Last opp profilbilde</span>
+								<span>Nytt profilbilde</span>
 								<input type="file">
-								{{ field }}
 							</div>
 							<div class="file-path-wrapper">
 								<input class="file-path validate" type="text">
@@ -93,26 +67,13 @@
 						{{ hidden }}
 						{% endfor %}
 					{% endfor %}
-				{% endfor %}
 			</div>
-					<div class="col s12 input-field" id="duty-section">
-						<h5>Obs! Vaktlisten på nettsiden er foreløpig placeholder og benyttes til testing av DevOps.<br><br>Vennligst benytt vaktlisten som ligger i Google Drive.</h5>	 
-					</div>
 			<input class="btn btn-large hs-green" type="submit" value="Submit"/>
 		</form>
 	</div>
 </div>
 </div>
 <script>
-	var vaktGetter = document.querySelector('#duty-section');
-	var autoDutyGetter = document.querySelector('#autocheckbox');
-	autoDutyGetter.addEventListener('click', function() {
-		if(vaktGetter.style.display === "none") {
-			vaktGetter.style.display = "block";
-		} else {
-			vaktGetter.style.display = "none";
-		}
-	});
 	M.AutoInit();
 </script>
 {% endblock %}
