+++ website/templates/website/header.html
@@ -1,5 +1,4 @@
 {% load staticfiles %}
-{% load loginmodal %}
 {% if request.user.is_authenticated and not request.user.profile.tos_accepted %}
 	{% if request.path == "/tos/returning-user/" or request.path == "/tos/" %}
 	{% else %}
@@ -74,21 +73,17 @@
 
 					<li><div class="divider"></div></li>
 					{% if user.is_authenticated %}
-					<li><a class="waves-effect" href="{% url 'auth:password_change' %}">Bytt Passord</a></li>
 					<li><a class="waves-effect" href="{% url 'auth:logout' %}">Logg Ut</a></li>
 					{% else %}
-					<li><a class="waves-effect" href="{% url 'auth:signup' %}">Registrer</a></li>
 					<li><a class="waves-effect" href="{% url 'auth:login' %}">Logg inn</a></li>
 					{% endif %}
 					<!-- Dropdown user list for desktop-->
 					<ul id="userdropdown" class="dropdown-content">
 						{% if user.is_authenticated %}
 						<li><a href="{% url 'userprofile:profile' %}">Profil</a></li>
-						<li><a href="{% url 'auth:password_change' %}">Bytt Passord</a></li>
 						<li><a href="{% url 'auth:logout' %}">Logg Ut</a></li>
 						{% else %}
-						<li><a href="{% url 'auth:signup' %}">Registrer</a></li>
-						<li><a class="modal-trigger" href="#loginModal">Logg Inn</a></li>
+						<li><a class="modal-trigger" href="{% url 'auth:login' %}">Logg Inn</a></li>
 						{% endif %}
 					</ul>
 					<!-- Dropdown for resurser -->
@@ -108,11 +103,6 @@
 			</div>
 		</nav>
 	</div>
-<div id="loginModal" class="modal">
-	<div class="modal-content">
-		{% LoginModal %}
-	</div>
-</div>
 <div id="mapModal" class="modal bottom-sheet">
 	<div class="modal-content">
 		<iframe frameborder="0" scrolling="no" marginheight="0" marginwidth="0" src="https://use.mazemap.com/embed.html#v=1&zlevel=1&campusid=1&sharepoitype=poi&sharepoi=71922" style="border: 1px solid grey; height:35vh; width:100%;"></iframe>
